+++ src/graph/expression_operators.h
@@ -105,9 +105,9 @@ Expr atleast_nd(Expr a, size_t dims);
 Expr flatten(Expr a);
 Expr flatten_2d(Expr a);
 
-Expr rows(Expr a, const std::vector<size_t>& indices);
-Expr cols(Expr a, const std::vector<size_t>& indices);
-Expr select(Expr a, int axis, const std::vector<size_t>& indices);
+Expr rows(Expr a, const std::vector<uint32_t>& indices);
+Expr cols(Expr a, const std::vector<uint32_t>& indices);
+Expr select(Expr a, int axis, const std::vector<uint32_t>& indices);
 
 /*********************************************************/
